INTCMN-141: Pulled inner class enums out of views
